+++ node/standalone/Cargo.toml
@@ -64,7 +64,6 @@ circuit-standalone-runtime = { path = "../../runtime/standalone" }
 pallet-3vm-contracts-rpc   = { path = "../../3vm/rpc/contracts", package = "pallet-contracts-rpc" }
 pallet-3vm-evm-rpc         = { path = "../../3vm/rpc/evm", package = "pallet-evm-rpc" }
 pallet-portal-rpc          = { path = "../../pallets/portal/rpc" }
-pallet-treasury            = { path = "../../pallets/treasury" }
 pallet-xdns                = { path = "../../pallets/xdns" }
 pallet-xdns-rpc            = { path = "../../pallets/xdns/rpc" }
 t3rn-primitives            = { path = "../../primitives" }
