+++ node/standalone/src/chain_spec.rs
@@ -368,7 +368,7 @@ fn testnet_genesis(
         },
         contracts_registry: Default::default(),
         account_manager: Default::default(),
-        treasury: Default::default(),
+        clock: Default::default(),
         three_vm: Default::default(), // TODO: genesis for this needs to be setup for the function pointers\
         evm: EvmConfig {
             // We need _some_ code inserted at the precompile address so that
