+++ node/t3rn-parachain/Cargo.toml
@@ -24,7 +24,6 @@ serde        = { version = "1.0", features = [ "derive" ] }
 circuit-parachain-runtime = { path = "../../runtime/t3rn-parachain", package = "t3rn-parachain-runtime" }
 pallet-3vm-contracts-rpc  = { path = "../../3vm/rpc/contracts", package = "pallet-contracts-rpc" }
 pallet-3vm-evm-rpc        = { path = "../../3vm/rpc/evm", package = "pallet-evm-rpc" }
-pallet-treasury           = { path = "../../pallets/treasury" }
 pallet-xdns               = { path = "../../pallets/xdns" } # TODO: the pallet shouldnt be imported in the node, expose primitives or export the type from runtime
 pallet-xdns-rpc           = { path = "../../pallets/xdns/rpc" }
 t3rn-primitives           = { path = "../../primitives" }
