+++ pallets/executors/src/subject_metadata.rs
@@ -16,12 +16,13 @@ use sp_runtime::{
 };
 use sp_std::prelude::*;
 use t3rn_primitives::{
+    // treasury::Treasury,
+    clock::Clock,
     common::{OrderedSet, RoundIndex},
     executors::{
         Bond, CandidateBondLessRequest, CapacityStatus, ExecutorStatus, StakeAdjust, StakerAdded,
         StakerStatus, STAKER_LOCK_ID,
     },
-    treasury::Treasury,
 };
 
 #[derive(Clone, Encode, Decode, RuntimeDebug, TypeInfo, PartialEq)]
