--- pallets/treasury/Cargo.toml
@@ -1,54 +1,0 @@
-[package]
-authors     = [ "t3rn devs <team@t3rn.io" ]
-description = "Implements initial inflation and rewards allocation."
-edition     = "2021"
-homepage    = "https://t3rn.io/"
-license     = "Apache 2.0"
-name        = "pallet-treasury"
-publish     = false
-repository  = "https://github.com/t3rn/t3rn"
-version     = "0.0.0"
-
-[package.metadata.docs.rs]
-targets = [ "x86_64-unknown-linux-gnu" ]
-
-[dependencies]
-codec      = { package = "parity-scale-codec", version = "3", default-features = false, features = [ "derive", "max-encoded-len" ] }
-log        = { version = "0.4", default-features = false }
-scale-info = { version = "2.1.1", default-features = false, features = [ "derive" ] }
-serde      = { version = "1.0", default-features = false, features = [ "derive" ] }
-
-## substrate deps
-frame-benchmarking = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.27", default-features = false }
-frame-support      = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.27", default-features = false }
-frame-system       = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.27", default-features = false }
-sp-runtime         = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.27", default-features = false }
-
-pallet-balances = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.27", default-features = false }
-pallet-staking  = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.27", default-features = false }
-sp-std          = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.27", default-features = false }
-t3rn-primitives = { path = "../../primitives", default-features = false }
-
-fixed = { tag = "v0.5.9", git = "https://github.com/encointer/substrate-fixed", package = "substrate-fixed", default-features = false }
-
-[dev-dependencies]
-similar-asserts = "1.1.0"
-sp-core         = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.27" }
-sp-io           = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.27" }
-
-[features]
-default = [ "std" ]
-std = [
-  "codec/std",
-  "serde/std",
-  "scale-info/std",
-  "frame-support/std",
-  "frame-system/std",
-  "frame-benchmarking/std",
-  "pallet-balances/std",
-  "pallet-staking/std",
-  "t3rn-primitives/std",
-]
-
-runtime-benchmarks = [ "frame-benchmarking/runtime-benchmarks" ]
-try-runtime        = [ "frame-support/try-runtime" ]
