--- pallets/treasury/README.md
@@ -1,3 +1,0 @@
-# pallet-treasury
-
-A custom treasury pallet covering a.o. initial inflation and rewards allocation.
