--- primitives/src/treasury.rs
@@ -1,5 +1,0 @@
-use crate::common::RoundInfo;
-
-pub trait Treasury<T: frame_system::Config> {
-    fn current_round() -> RoundInfo<T::BlockNumber>;
-}
