+++ runtime/common-pallets/Cargo.toml
@@ -82,7 +82,6 @@ pallet-collator-selection           = { git = 'https://github.com/paritytech/cum
 parachain-info                      = { git = 'https://github.com/paritytech/cumulus', branch = 'polkadot-v0.9.27', default-features = false }
 
 # t3rn Deps
-pallet-treasury = { default-features = false, path = "../../pallets/treasury" }
 t3rn-primitives = { default-features = false, path = "../../primitives" }
 t3rn-protocol   = { default-features = false, path = "../../protocol" }
 
@@ -182,7 +181,6 @@ std = [
   # t3rn pallets
   "pallet-xdns/std",
   "pallet-xdns-rpc-runtime-api/std",
-  "pallet-treasury/std",
   "pallet-contracts-registry/std",
   "pallet-circuit/std",
   "pallet-grandpa/std",
@@ -204,7 +202,6 @@ runtime-benchmarks = [
   "sp-runtime/runtime-benchmarks",
   "xcm-builder/runtime-benchmarks",
   "cumulus-pallet-session-benchmarking/runtime-benchmarks",
-  "pallet-treasury/runtime-benchmarks",
 ]
 
 try-runtime = [ "frame-executive/try-runtime", "frame-try-runtime" ]
