+++ runtime/common-pallets/src/lib.rs
@@ -22,7 +22,6 @@ pub use pallet_clock;
 pub use pallet_contracts_registry;
 pub use pallet_grandpa_finality_verifier;
 pub use pallet_portal;
-pub use pallet_treasury;
 pub use pallet_xbi_portal;
 pub use pallet_xbi_portal_enter;
 pub use pallet_xdns;
