+++ runtime/mock/src/lib.rs
@@ -55,7 +55,7 @@ frame_support::construct_runtime!(
         ContractsRegistry: pallet_contracts_registry::{Pallet, Call, Config<T>, Storage, Event<T>} = 106,
         Circuit: pallet_circuit::{Pallet, Call, Storage, Event<T>} = 108,
         Treasury: pallet_treasury = 109,
-        Clock: pallet_clock::{Pallet, Storage, Event<T>} = 110,
+        Clock: pallet_clock::{Pallet, Config<T>, Storage, Event<T>} = 110,
 
         XBIPortal: pallet_xbi_portal::{Pallet, Call, Storage, Event<T>} = 111,
         XBIPortalEnter: pallet_xbi_portal_enter::{Pallet, Call, Event<T>} = 112,
