+++ runtime/standalone/src/circuit_config.rs
@@ -18,7 +18,6 @@ impl pallet_clock::Config for Runtime {
     type Event = Event;
     type Executors = t3rn_primitives::executors::ExecutorsMock<Self>;
     type RoundDuration = ConstU32<500u32>;
-    type Treasury = Treasury;
 }
 
 impl pallet_xdns::Config for Runtime {
@@ -97,35 +96,4 @@ impl pallet_grandpa_finality_verifier::Config<RococoBridgeInstance> for Runtime 
     type BridgedChain = Blake2ValU32Chain;
     type HeadersToStore = HeadersToStore;
     type WeightInfo = ();
-}
-
-// MinRoundTerm plays a crucial role:
-//  + must at least be the size of the active collator set
-//  + is applied as default round term during genesis
-//  + codetermines staking delays as they are measured in rounds
-parameter_types! {
-    pub const TreasuryAccount: AccountId = AccountId::new([0u8; 32]); // TODO
-    pub const ReserveAccount: AccountId = AccountId::new([1u8; 32]); // TODO
-    pub const AuctionFund: AccountId = AccountId::new([2u8; 32]); // TODO
-    pub const ContractFund: AccountId = AccountId::new([3u8; 32]); // TODO
-    pub const MinRoundTerm: u32 = 20; // TODO
-    pub const DefaultRoundTerm: u32 = DEFAULT_ROUND_TERM; // TODO
-    pub const GenesisIssuance: u32 = 20_000_000; // TODO
-    pub const IdealPerpetualInflation: Perbill =Perbill::from_percent(1);
-    pub const InflationRegressionMonths: u32 = 72;
-}
-
-impl pallet_treasury::Config for Runtime {
-    type AuctionFund = AuctionFund;
-    type ContractFund = ContractFund;
-    type Currency = Balances;
-    type DefaultRoundTerm = DefaultRoundTerm;
-    type Event = Event;
-    type GenesisIssuance = GenesisIssuance;
-    type IdealPerpetualInflation = IdealPerpetualInflation;
-    type InflationRegressionMonths = InflationRegressionMonths;
-    type MinRoundTerm = MinRoundTerm;
-    type ReserveAccount = ReserveAccount;
-    type TreasuryAccount = TreasuryAccount;
-    type WeightInfo = pallet_treasury::weights::TreasuryWeight<Runtime>;
 }
